Added visual effect for signal sending;
fixed signal looping between two elements

+++ src/main/java/com/fi0x/cc/project/mixer/TimeCalculator.java
@@ -2,7 +2,7 @@ package com.fi0x.cc.project.mixer;
 
 public class TimeCalculator
 {
-    public static int getMillisFromBeat(int noteCount)
+    public static int getMillisFromBeat(float noteCount)
     {
         float beat = 60000f / MixerManager.getBPM();
         float beatParts = beat / MixerManager.getNotesPerBeat();
